refactor(adoc-core): unit 6 — split SourceProvider trait into domain/ports, FsSourceProvider/InMemorySourceProvider into infrastructure/source
git mv src/source_provider.rs src/infrastructure/source/fs.rs. Remove
SourceProvider trait + SourceLoadError struct from fs.rs (now in
domain/ports/source_provider.rs from Unit 4). Cut InMemorySourceProvider
(with #[cfg(test)] gate intact) into new infrastructure/source/in_memory.rs.
Create infrastructure/source/mod.rs. Update compile.rs to use new paths.
Remove mod source_provider from lib.rs.

Refs #27.
