+++ crates/adoc-core/src/infrastructure/mod.rs
@@ -1,1 +1,2 @@
 // see ADR-0009
+pub(crate) mod source;
