+++ crates/adoc-core/src/infrastructure/source/mod.rs
@@ -1,0 +1,3 @@
+// see ADR-0009
+pub(crate) mod fs;
+pub(crate) mod in_memory;
